Add ContextWrapper
